complete satisfies process
### Remaining Work
Need to verify that all license and license ref options work.  May need to add processing for some aspects of licenses and license refs.  Some regression tests continue to fail.  All are marked with a comment about the failure.

Includes refactor to use *Node instead of string when processing satisfies.