allow options to be passed in as nil
There is not a substantial difference between the approach of requiring the options parameter to be an empty struct (i.e. `Options{}`) vs `nil`.

Allowing `nil` provides a convenience for the caller.  The cost of allowing `nil` is an additional check and creation of the empty `Options` struct for every call.