rename spdx package expression
With the rename of spdx-expressions repo to go-spdx, the name of the package reads better as expression.  This allows for future changes to the API be something like…

go-spdx/expression - validation of expressions
go-spdx/license - validation of licenses
go-spdx/parse - parsing

This naming scheme more closely aligns with go conventions.